+++ packages/vsdx/lib/vsdx.dart
@@ -67,6 +67,7 @@ export 'src/parser/parse_visio.dart';
 export 'src/parser/vdx/vdx_document_parser.dart';
 export 'src/parser/emf_embedded_bitmap.dart';
 export 'src/parser/metafile.dart';
+export 'src/parser/metafile_png.dart';
 export 'src/parser/vsd/vsd_document_parser.dart';
 export 'src/parser/vsd/cfb/compound_file.dart';